requirementsPython.txt
mysql-connector-python==8.0.19
pypcapfile==0.12.0
requests==2.22.0
requests-oauthlib==1.3.0